Transform invalid tool usages on sending, not on initial detection (#1091)
Per bug #1069, session-managers never persist tool-name changes after we initially persist the message, which means once an agent generates an invalid-tool name, that message history is poisoned  on re-hydration.  To avoid that going forward, do the translation of invalid-tool names on sending to the provider and not on the initial tool_use detection.  The initial tool_use detection is needed to add a tool_response with a proper error message for the LLM, but this will avoid the poisoning issue

---------

Co-authored-by: Mackenzie Zastrow <zastrowm@users.noreply.github.com>