fix: Don't hang when MCP server returns 5xx (#1169)
Fixes #995 where if a MCP tool_call receives a 5XX error from the server, the call hangs and never ends. The root cause is that Anthropic's MCP client - on receiving a 5XX - bubbles up an exception that ends up cancelling all TaskGroup tasks which results in the session/client/asyncio loop being torn down and the tool_call never resolves, thus the hang.

The fix is two fold:

- Detect that the situation occurs and trigger a close `close_future` future
- Update all background_invokes to eagerly bail on `close_future` being triggered

---------

Co-authored-by: Mackenzie Zastrow <zastrowm@users.noreply.github.com>