fix: Pass CODECOV_TOKENS through for code-coverage stats (#1385)
We need to plumb this through so that protected branches (like main) correctly report code coverage to codecov.  Prior to this PRs did not have incremental diffs because main builds didn't have this token; only `main` needs the token because it's a protected branch [0].

This token is already a secret in this repository.

[0]: https://docs.codecov.com/docs/codecov-tokens#when-do-i-need-a-token

Co-authored-by: Mackenzie Zastrow <zastrowm@users.noreply.github.com>